Base frame for new executor (#354)
