+++ yajsapi/agreement/index.ts
@@ -1,3 +1,4 @@
 import { AgreementPoolService } from "./agreement_pool_service";
+import { Agreement } from "./agreement";
 
-export { AgreementPoolService };
+export { AgreementPoolService, Agreement };
