+++ yajsapi/payment/payment_service.ts
@@ -8,22 +8,11 @@ export class PaymentService {
     private readonly logger?: Logger
   ) {}
   async run() {
-    this.process_debit_notes();
-    this.process_invoices();
+    this.logger?.info("The Payment Service has started");
   }
-  acceptPayments(agreementId: string) {
-    this.accept_payment_for_agreement({ agreement_id: agreementId, partial: false });
-  }
-  // TODO: old references - to refactor
-  async accept_payment_for_agreement({ agreement_id, partial }) {
-    // only reference mocked
-  }
-  async process_debit_notes() {
-    // only reference mocked
-  }
-  async process_invoices() {
-    // only reference mocked
-  }
+  acceptPayments(agreementId: string) {}
 
-  async end() {}
+  async end() {
+    // todo
+  }
 }
